Drawing HUD completely in C++
Create a C++ Class that inherits from HUD call it "DebugHUD" then create a C++ class that inherits from "DebugHUD" call it "BallBearingHUD".
This entire code is just a boilerplate to allow us to draw a text to the viewport in a consistent way.

We created 2 C++ HUD classes cause we might wanna add another Pawn with different parameters.

note that we can't access  APlayerBallBearing* without making ABallBearingHUD a friend class at PlayerBallBearing C++ header.

Remember that StaticClass() returns a pointer of the UClass type." representing this class at runtime". a static class cannot be instantiated. In other words, you cannot use the new operator to create a variable of the class type.
What we are saying to the engine is-> create yourself an object from this particular HUD when the game starts.