Resetting location after falling with TriggerBox Blueprint
Moved BP_PlayerBallBearing to a new folder within called "BallBearings" .. and then created a BP class that inherits from TriggerBox called it "BP_ResetBallBearings". and inside from Even ActorBeginOverlap we cast to BallBearing and ResetLocation.
and then Drag the "BP_ResetBallBearings" into the level and scale it to cover beneath the floor wherever BallBearing falls.