Setting up Goals
Created a LevelBallBearing Blueprint class based on BallBearing C++ Class and changed the mass of it to differentiate from BP_PlayerBallBearing 25 Kg (Not 100). and changed it material to differentiate as well from BP_PlayerBallBearing material.

Then we created A C++ class for BallBearingGoal that inherits from Trigger Sphere, and then created a BP class that inherits from it called it BP_BallBearingGoal.

Then we head to the BP_BallBearingGoal and Add Component-> Static Mesh and add a "Torus" to shape the goal, and then Add Component-> Particle System and set Particle Template to "Fire" or something flashy.

Then we add the Goals to our level and LevelBallBearings and try to match Balls into Goals

Note that:
PostInitializeComponents() Allows actors to initialize themselves on the C++ side after all of their components have been initialized, only called during gameplay. Instead of loading on begin play and limit the traffic on the first frame of the game.